refactor(components/molecule/select): api tab
